+++ src/drivers/bt/pit.cpp
@@ -59,7 +59,7 @@ Pit::Pit(tSituation *s, Driver *driver)
 		// Fix broken pit exit.
 		if (p[6].x < p[5].x) {
 			//printf("bt: Pitexit broken on track %s.\n", track->name);
-			p[6].x = p[5].x + 50.0;
+			p[6].x = p[5].x + 50.0f;
 		}
 
 		// Fix point for first pit if necessary.
@@ -72,7 +72,7 @@ Pit::Pit(tSituation *s, Driver *driver)
 			p[5].x = p[4].x;
 		}
 
-		float sign = (pitinfo->side == TR_LFT) ? 1.0 : -1.0;
+		float sign = (float)((pitinfo->side == TR_LFT) ? 1.0 : -1.0);
 		p[0].y = 0.0;
 		p[6].y = 0.0;
 		for (i = 1; i < NPOINTS - 1; i++) {
